changes from code review
